+++ np.CallbackURLs/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 See Plugin [README](https://github.com/NotePlan/plugins/blob/main/np.CallbackURLs/README.md) for details on available commands and use cases.
 
+## [1.6.2] - 2024-05-28 @dwertheimer
+
+- Bump version to re-release it
+
 ## [1.6.1] - 2023-09-12 @dwertheimer
 
 - Bug fix for calling np.Templating (can't pull plugin.json)